+++ streampark-console/streampark-console-service/src/main/java/org/apache/streampark/console/api/v2/controller/AlertConfigsApiImpl.java
@@ -22,8 +22,8 @@ import org.apache.streampark.console.api.controller.model.AlertConfigInfo;
 import org.apache.streampark.console.api.controller.model.CreateAlertConfigRequest;
 import org.apache.streampark.console.api.controller.model.ListAlertConfig;
 import org.apache.streampark.console.api.controller.model.UpdateAlertConfigRequest;
-import org.apache.streampark.console.api.exception.NotFoundException;
 import org.apache.streampark.console.api.v2.convert.AlertConfigConvert;
+import org.apache.streampark.console.api.v2.exception.NotFoundException;
 import org.apache.streampark.console.core.entity.AlertConfig;
 import org.apache.streampark.console.core.service.alert.AlertConfigService;
 import org.apache.streampark.console.core.service.alert.AlertService;
